requirements.txt
streamlit==1.29.0
pandas==2.1.4 
openai==1.58.1
python-dotenv==1.0.1
boto3==1.35.40